+++ metadata.rb
@@ -11,14 +11,21 @@ git_url = 'https://github.com/ualaska-it/http_platform'
 source_url git_url if respond_to?(:source_url)
 issues_url "#{git_url}/issues" if respond_to?(:issues_url)
 
-version '1.4.4'
+version '2.0.0'
 
 supports 'ubuntu', '>= 16.0'
-supports 'centos', '>= 7.0'
+supports 'debian', '>= 9.0'
+supports 'redhat', '>= 6.0'
+supports 'centos', '>= 6.0'
+supports 'oracle', '>= 6.0'
+# supports 'fedora'
+# supports 'amazon'
+# supports 'suse'
+# supports 'opensuse'
 
 chef_version '>= 14.0' if respond_to?(:chef_version)
 
-depends 'apache2', '~> 5.2' # 7 support coming soon
+depends 'apache2', '>= 6.0'
 depends 'chef-vault'
 depends 'firewall'
 depends 'yum-epel'