+++ test/integration/default/cert_test.rb
@@ -4,6 +4,13 @@ require_relative '../helpers'
 
 node = json('/opt/chef/run_record/last_chef_run_node.json')['automatic']
 
+owner_group =
+  if node['platform_family'] == 'debian'
+    'ssl-cert'
+  else
+    'root'
+  end
+
 alt_regex = 'subject_alt_name: \["DNS:www.funny.business", "DNS:funny.business", "DNS:www.me.also", "DNS:me.also"\]'
 # rubocop:disable Metrics/LineLength
 # alt_regex = 'subject_alt_name: \["DNS:www.funny.business", "DNS:funny.business", "DNS:www.localhost", "DNS:localhost", "DNS:www.me.also", "DNS:me.also"\]'
@@ -24,6 +31,7 @@ describe file('/opt/chef/run_record/http_cert_record.txt') do
   its(:content) { should match 'org_unit: fake_unit' }
   its(:content) { should match 'email: fake-it@make-it' }
   its(:content) { should match 'expire: 365' }
+  its(:content) { should match "group: #{owner_group}" }
   its(:content) { should match 'key_type: \'rsa\'' }
   its(:content) { should match 'key_length: 2048' }
 end
@@ -34,7 +42,25 @@ describe file('/opt/chef/run_record/http_key_record.txt') do
   it { should be_mode 0o644 }
   it { should be_owned_by 'root' }
   it { should be_grouped_into 'root' }
+  its(:content) { should match "group: #{owner_group}" }
   its(:content) { should match 'key_length: 2048' }
+end
+
+describe file(path_to_self_signed_key(node)) do
+  it { should exist }
+  it { should be_file }
+  it { should be_mode 0o640 }
+  it { should be_owned_by 'root' }
+  it { should be_grouped_into key_group(node) }
+  its(:content) { should match '-----BEGIN RSA PRIVATE KEY-----\nMIIE' }
+end
+
+describe key_rsa(path_to_self_signed_key(node)) do
+  it { should be_public }
+  it { should be_private }
+  its('public_key') { should match '-----BEGIN PUBLIC KEY-----\nMIIB' }
+  its('private_key') { should match '-----BEGIN RSA PRIVATE KEY-----\nMIIE' }
+  its('key_length') { should eq 2048 }
 end
 
 describe file(path_to_self_signed_cert(node)) do
@@ -44,33 +70,6 @@ describe file(path_to_self_signed_cert(node)) do
   it { should be_owned_by 'root' }
   it { should be_grouped_into 'root' }
   its(:content) { should match 'BEGIN CERTIFICATE' }
-end
-
-describe file(path_to_self_signed_key(node)) do
-  it { should exist }
-  it { should be_file }
-  it { should be_mode 0o640 }
-  it { should be_owned_by 'root' }
-  it { should be_grouped_into key_group(node) }
-  its(:content) { should match 'BEGIN RSA PRIVATE KEY' }
-end
-
-describe file(path_to_dh_config(node)) do
-  it { should exist }
-  it { should be_file }
-  it { should be_mode 0o644 }
-  it { should be_owned_by 'root' }
-  it { should be_grouped_into 'root' }
-  its(:content) { should match '2048' }
-end
-
-describe file(path_to_dh_params(node)) do
-  it { should exist }
-  it { should be_file }
-  it { should be_mode 0o644 }
-  it { should be_owned_by 'root' }
-  it { should be_grouped_into 'root' }
-  its(:content) { should match 'BEGIN DH PARAMETERS' }
 end
 
 describe x509_certificate(path_to_self_signed_cert(node)) do
@@ -103,6 +102,24 @@ describe x509_certificate(path_to_self_signed_cert(node)) do
   its('issuer.L') { should eq 'Fairbanks' }
   its('issuer.O') { should eq 'fake_org' }
   its('issuer.OU') { should eq 'fake_unit' }
+end
+
+describe file(path_to_dh_config(node)) do
+  it { should exist }
+  it { should be_file }
+  it { should be_mode 0o644 }
+  it { should be_owned_by 'root' }
+  it { should be_grouped_into 'root' }
+  its(:content) { should match '2048' }
+end
+
+describe file(path_to_dh_params(node)) do
+  it { should exist }
+  it { should be_file }
+  it { should be_mode 0o644 }
+  it { should be_owned_by 'root' }
+  it { should be_grouped_into 'root' }
+  its(:content) { should match 'BEGIN DH PARAMETERS' }
 end
 
 describe file(File.join(cert_public_dir(node), 'http_platform_csr.pem')) do