+++ test/integration/lets_encrypt_cert/lets_encrypt_test.rb
@@ -4,6 +4,23 @@ require_relative '../helpers'
 
 node = json('/opt/chef/run_record/last_chef_run_node.json')['automatic']
 
+describe file(path_to_lets_encrypt_key) do
+  it { should exist }
+  it { should be_file }
+  it { should be_mode 0o640 }
+  it { should be_owned_by 'root' }
+  it { should be_grouped_into key_group(node) }
+  its(:content) { should match '-----BEGIN RSA PRIVATE KEY-----\nMIIE' }
+end
+
+describe key_rsa(path_to_lets_encrypt_key) do
+  it { should be_public }
+  it { should be_private }
+  its('public_key') { should match '-----BEGIN PUBLIC KEY-----\nMIIB' }
+  its('private_key') { should match '-----BEGIN RSA PRIVATE KEY-----\nMIIE' }
+  its('key_length') { should eq 2048 }
+end
+
 describe file(path_to_lets_encrypt_cert) do
   it { should exist }
   it { should be_file }
@@ -13,13 +30,25 @@ describe file(path_to_lets_encrypt_cert) do
   its(:content) { should match 'BEGIN CERTIFICATE' }
 end
 
-describe file(path_to_lets_encrypt_key) do
-  it { should exist }
-  it { should be_file }
-  it { should be_mode 0o640 }
-  it { should be_owned_by 'root' }
-  it { should be_grouped_into key_group(node) }
-  its(:content) { should match 'BEGIN RSA PRIVATE KEY' }
+describe x509_certificate(path_to_self_signed_cert(node)) do
+  its('version') { should eq 2 }
+  its('key_length') { should eq 2048 }
+  its('signature_algorithm') { should eq 'sha256WithRSAEncryption' }
+  its('validity_in_days') { should be > 89 }
+  its('validity_in_days') { should be < 91 }
+
+  its('extensions') { should include 'subjectAltName' }
+
+  # Need to update these
+  its('issuer.CN') { should eq 'Let\'s Encrypt Authority X3' }
+  its('issuer.emailAddress') { should eq '???t' }
+  its('issuer.C') { should eq '??' }
+  its('issuer.ST') { should eq '??' }
+  its('issuer.L') { should eq '??' }
+  its('issuer.O') { should eq 'Let\'s Encrypt' }
+  its('issuer.OU') { should eq '??' }
+
+  its('issuer_dn') { should match 'CN=??' }
 end
 
 describe apache_conf(File.join(path_to_conf_available_dir(node), 'ssl-params.conf')) do